refactor(agent): blueprints, phase helpers, named constants (review)
Address the module-level review feedback:

- server/routes.py → server/routes/ package of Flask blueprints (brand,
  system, tools, chat) with a register_all(app) helper; main.py registers them.
- chat_stream's per-phase logic extracted into workflows/chat_pipeline.py
  (run_mcp_phase / run_kb_phase / run_synthesis_phase / run_followups) that the
  route delegates to via `yield from`. SSE payloads and phase order unchanged.
- session_logger: remove the no-op _truncate_response; name the response-preview
  length (MAX_TEXT_PREVIEW_LENGTH).
- Name magic numbers: SYNTHESIS_PREVIEW_LENGTH (chart_config),
  MAX_FOLLOW_UP_QUESTIONS (follow_up).
- Make values config-driven: KB temperature/dynamicThreshold; MCP handshake
  clientInfo/protocolVersion. Add a TODO for the query_param_key default.
- Bump Copyright 2024 -> 2026 across the agent files.
