+++ narratives/agent/src/server/routes/system.py
@@ -1,0 +1,82 @@
+#!/usr/bin/env python3
+# Copyright 2026 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from flask import Blueprint, jsonify
+
+from src.config import load_config
+from src.mcp.client import MCP_PORT, MCP_URL
+from src.server.app import PROXY_PORT
+
+system_bp = Blueprint("system", __name__)
+
+
+# Flask Routes
+
+@system_bp.route("/health", methods=["GET"])
+def health():
+    """Health check."""
+    return jsonify({"status": "ok", "mcp_url": MCP_URL})
+
+
+@system_bp.route("/", methods=["GET"])
+def index():
+    return f"""
+    <html>
+    <head><title>MCP Proxy</title></head>
+    <body>
+    <h1>Data Commons MCP Proxy Server (Proxy-Only Mode)</h1>
+    <p>MCP Server: {MCP_URL}</p>
+    <p>Proxy Server: http://localhost:{PROXY_PORT}</p>
+    <ul>
+        <li><a href="/health">/health</a> - Health check</li>
+        <li><a href="/api/tools">/api/tools</a> - List tools</li>
+        <li>POST /api/call - Execute tool</li>
+        <li><a href="/api/config">/api/config</a> - Get backend config (no API key)</li>
+        <li>POST /api/chat/stream - Full chat with streaming</li>
+        <li><a href="/logs?key=">/logs</a> - Query Analytics Dashboard (requires ?key=SECRET)</li>
+    </ul>
+    <h3>Prerequisite</h3>
+    <p>Make sure the MCP server is running:</p>
+    <code>python3 -m uv tool run datacommons-mcp serve http --port {MCP_PORT}</code>
+    </body>
+    </html>
+    """
+
+
+# ============================================================
+# NEW BACKEND API ENDPOINTS FOR GEMINI CALLS
+# ============================================================
+
+@system_bp.route("/api/config", methods=["GET"])
+def get_config_endpoint():
+    """Return sanitized config (without API key) for frontend."""
+    config = load_config()
+    if not config:
+        return jsonify({"success": False, "error": "Config not loaded"}), 500
+
+    # Return config without sensitive data
+    safe_config = {
+        "proxy_url": config.get("proxy_url", f"http://localhost:{PROXY_PORT}"),
+        "gemini": {
+            "api_base": config.get("gemini", {}).get("api_base", ""),
+            "mcp_model": config.get("gemini", {}).get("mcp_model", "gemini-3-flash-preview"),
+            "kb_model": config.get("gemini", {}).get("kb_model", "gemini-3-flash-preview"),
+        },
+        "mcp": config.get("mcp", {}),
+        "knowledge_base": config.get("knowledge_base", {}),
+        "thinking": config.get("thinking", {}),
+        "has_api_key": bool(config.get("gemini", {}).get("api_key")),
+    }
+    return jsonify({"success": True, "config": safe_config})
