+++ narratives/agent/src/server/routes/tools.py
@@ -1,0 +1,65 @@
+#!/usr/bin/env python3
+# Copyright 2026 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import logging
+
+from flask import Blueprint, jsonify, request
+
+import src.mcp.client as mcp_client
+from src.mcp.client import call_tool, get_tools, initialize_mcp
+from src.mcp.schema import transform_schema_for_gemini
+
+logger = logging.getLogger(__name__)
+
+tools_bp = Blueprint("tools", __name__)
+
+
+@tools_bp.route("/api/tools", methods=["GET"])
+def list_tools():
+    """List available tools."""
+    if not mcp_client.session_id:
+        if not initialize_mcp():
+            return jsonify({"success": False, "error": "Cannot connect to MCP server. Make sure it's running on port 3000!"}), 503
+
+    tools = get_tools()
+    if not tools:
+        return jsonify({"success": False, "error": "No tools available"}), 503
+
+    # Convert to Gemini format (transform schema to remove unsupported constructs)
+    gemini_tools = [{
+        "name": t.get("name", ""),
+        "description": t.get("description", ""),
+        "parameters": transform_schema_for_gemini(
+            t.get("inputSchema", {"type": "object", "properties": {}})
+        )
+    } for t in tools]
+
+    return jsonify({"success": True, "tools": gemini_tools, "raw_tools": tools})
+
+
+@tools_bp.route("/api/call", methods=["POST"])
+def tool_call():
+    """Execute a tool call."""
+    if not mcp_client.session_id:
+        if not initialize_mcp():
+            return jsonify({"success": False, "error": "Cannot connect to MCP server"}), 503
+
+    data = request.get_json()
+    if not data or not data.get("name"):
+        return jsonify({"success": False, "error": "Tool name required"}), 400
+
+    logger.info(f"Calling tool: {data['name']}")
+    result = call_tool(data["name"], data.get("arguments", {}))
+    return jsonify({"success": True, "result": result})
