+++ narratives/agent/src/workflows/follow_up.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright 2024 Google LLC
+# Copyright 2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,6 +22,9 @@ from src.gemini.client import gemini_request
 from src.gemini.schemas import DEFAULT_FOLLOW_UP_PROMPT, FOLLOW_UP_SCHEMA
 
 logger = logging.getLogger(__name__)
+
+# Max number of follow-up questions returned to the UI.
+MAX_FOLLOW_UP_QUESTIONS = 3
 
 
 def generate_follow_up_questions(user_message: str, topics: list) -> list:
@@ -76,6 +79,6 @@ Generate the self-contained follow-up questions now."""
             continue
         seen.add(key)
         cleaned.append(q)
-        if len(cleaned) >= 3:
+        if len(cleaned) >= MAX_FOLLOW_UP_QUESTIONS:
             break
     return cleaned
